完成了地图上标记出信号强度，android通过okhttp上传数据到服务器，服务端存数据到mysql，前端每隔10s取数据标记

+++ pycharm/djangoProject/app/models.py
@@ -11,7 +11,7 @@ class Register(models.Model):
 class LocationData(models.Model):
     latitude = models.FloatField()
     longitude = models.FloatField()
-    signal_strength = models.IntegerField()
+    signal_strength = models.FloatField()
     is_sent = models.BooleanField(default=False)
     is_ask = models.BooleanField(default=False)
